Use the Vector Base Address Register.
This chang euses the Vector Base Address Register (available in ARMv6
with Secure Extensions) to set the interrupt vector to something other
than 0x0 or the high-vector address.

Please note that there is a bug in QEMU which means that this register
is not available on the arm1176. Moving the VBAR CP15 reg to the armv6
definition seems to fix it, but I have yet to make a patch.