Turn interrupts on from kernel_main.
This change keeps IRQ/FIQ disabled until `kernel_main()` can mask the
BCM2835 interrupts. We do this because this code needs access to the
INT_DISABLE_* IO regs, which we will only have access to once
`mmu_init()` has run. (That isn't true yet, but it will be when a
subsequent change simplifies the initial MMU mapping).